fixing the code

+++ src/main/java/com/example/university/service/StudentJpaService.java
@@ -1,5 +1,6 @@
 package com.example.university.service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import com.example.university.model.*;
@@ -41,31 +42,82 @@ public class StudentJpaService implements StudentRepository {
 
     @Override
     public Student addStudent(Student student) {
+        if(studentJpaRepository.findById(student.getStudentId()).isPresent()){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        }
+        student.getCourses().forEach(course -> {
+            if(!courseJpaRepository.findById(course.getCourseId()).isPresent()){
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+            }
+        });
+//        List<Integer> courseIds = new ArrayList<>();
+//        student.getCourses().forEach(course -> {
+//           courseIds.add(course.getCourseId());
+//        });
+//        List<Course> courses = new ArrayList<>();
+//        courseIds.forEach(courseId ->{
+//            courses.add(courseJpaRepository.findById(courseId).get());
+//        });
+//        student.setCourses(courses);
         studentJpaRepository.save(student);
-        return student;
+        return studentJpaRepository.findById(student.getStudentId()).get();
     
     }
 
     @Override 
     public Student updateStudent(int studentId, Student student) {
+        if (!studentJpaRepository.findById(studentId).isPresent()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
         try {
             Student newStudent = studentJpaRepository.findById(studentId).get();
+            if(student.getCourses().size()>0){
+                List<Course> courses = student.getCourses();
+                courses.forEach(course -> {
+                    if(!courseJpaRepository.findById(course.getCourseId()).isPresent()){
+                        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+                    }
+                });
+                newStudent.setCourses(courses);
+            }
             if (student.getStudentName() != null)
                 newStudent.setStudentName(student.getStudentName());
             if (student.getEmail() != null)
                 newStudent.setEmail(student.getEmail());
            studentJpaRepository.save(newStudent);
            return newStudent;
-         } catch (Exception e) {
+         } catch (ResponseStatusException e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+         } catch (Exception e){
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-         }
+
+        }
     }
 
     @Override
     public void deleteStudent(int studentId) {
+        if(!studentJpaRepository.findById(studentId).isPresent()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
         try {
+            List<Course> updatedCourses = new ArrayList<>();
+            courseJpaRepository.findAll().forEach(course -> {
+                List<Student> newStudentList = new ArrayList<>();
+                course.getStudents().forEach(student -> {
+                    if (student.getStudentId() != studentId){
+                        newStudentList.add(student);
+                    }
+                });
+                course.setStudents(newStudentList);
+
+            });
             studentJpaRepository.deleteById(studentId);
-        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT);
+
+        }catch (ResponseStatusException re){
+            throw new ResponseStatusException(HttpStatus.NO_CONTENT);
+        }
+        catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
     }
